move CaseSyncUpdate to shared module
