+++ corehq/ex-submodules/casexml/apps/phone/data_providers/case/utils.py
@@ -1,0 +1,26 @@
+from casexml.apps.case import const
+
+
+class CaseSyncUpdate(object):
+    """
+    The record of how a case should sync
+    """
+    def __init__(self, case, sync_token, required_updates=None):
+        self.case = case
+        self.sync_token = sync_token
+        # cache this property since computing it can be expensive
+        self.required_updates = required_updates if required_updates is not None else self._get_required_updates()
+
+    def _get_required_updates(self):
+        """
+        Returns a list of the required updates for this case/token
+        pairing. Should be a list of actions like [create, update, close]
+        """
+        ret = []
+        if not self.sync_token or not self.sync_token.phone_has_case(self.case.get_id):
+            ret.append(const.CASE_ACTION_CREATE)
+        # always include an update
+        ret.append(const.CASE_ACTION_UPDATE)
+        if self.case.closed:
+            ret.append(const.CASE_ACTION_CLOSE)
+        return ret
